requirements.txt
aiogram==3.7.0
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.0
python-dotenv==1.0.1

# PostgreSQL / SQLAlchemy
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# HTTP / serialization
orjson==3.10.3
httpx==0.27.0
openpyxl==3.1.5

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
